requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
filelock==3.12.3
gdown==4.7.1
idna==3.4
Pillow==10.0.0
PySocks==1.7.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
